README
Target Codevita 2020 :  Day 1

Description: 

These are solutions to the coding problems of the first challenge. Problems are taken from Hackerrank and TCS Codevita challenge. All the codes are compiled at Hackerrank editor and coded in the form of functions. Most of the solutions are built in an easy manner, so that new learners can also understand the code. Codes are well commented.

The link to the problems are given below.   

Happy Coding.

https://www.hackerrank.com/contests/target-codevita-1/challenges/the-birthday-bar
https://www.hackerrank.com/contests/target-codevita-1/challenges/divisible-sum-pairs
https://www.hackerrank.com/contests/target-codevita-1/challenges/migratory-birds